% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfibos.R
\name{fibosfrom}
\alias{fibosfrom}
\title{subset fibonacci}
\usage{
fibosfrom(f0n = 0, fend = 15)
}
\arguments{
\item{f0n}{starting index}

\item{fend}{the index of the last fibonacci, starting from 0}
}
\value{
a numeric arrary that represents
a fibbonacci sequence from f0n to fend
}
\description{
subset fibonacci
}
\examples{
fibosfrom()

}
