% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfibos.R
\name{nfibos}
\alias{nfibos}
\title{Until nth fibonacci sequence}
\usage{
nfibos(fend)
}
\arguments{
\item{fend}{the index of the last fibonacci, starting from 0}
}
\value{
a numeric arrary that represents a fibbonacci sequence from 0 to fend
}
\description{
Until nth fibonacci sequence
}
\examples{
nfibos(4)

}
